+++ examples/basic_example.py
@@ -16,14 +16,13 @@ def r_example():
     # Try some simple calculations
     y1 = R.eval("sin(100)")
     y2 = R.eval(t("cos({y1})"))
-    print(f"{R.get(y1), R.get(y2)=}")
+    print(f"{y1.fetch(), y2.fetch()=}")
 
     # Remote objects can be printed without getting them first
     R.print(y2)
 
     # Bring a dataframe over from Rlang to python
-    remote_df = R.eval("iris")
-    df = R.get(remote_df, "pandas")
+    df = R.eval("iris").fetch("pandas")
     print(df)